requirements.txt
django==2.1.2
django-jet
djangorestframework==3.9.0
django-rest-swagger==2.2.0